Rename `metrics.ts` to `useBrowserId.ts`
